requirements.txt
anyio==4.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
markdownify==0.11.6
pypng==0.20220715.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
python-telegram-bot==20.7
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
telebot==0.0.5
typing_extensions==4.9.0
urllib3==2.1.0
